Сделал строгое API рерайтера, для устранения неоднозначностей по использованию
