+++ src/test/java/com/github/_1c_syntax/bsl/languageserver/utils/ExpressionParseTreeRewriterTest.java
@@ -319,9 +319,6 @@ class ExpressionParseTreeRewriterTest {
 
   BslExpression getExpressionTree(String code) {
     var expression = parse(code);
-    var rewriter = new ExpressionParseTreeRewriter();
-    expression.accept(rewriter);
-
-    return rewriter.getExpressionTree();
+    return ExpressionParseTreeRewriter.buildExpressionTree(expression);
   }
 }
