+++ src/test/java/com/github/_1c_syntax/bsl/languageserver/utils/ExpressionTreeComparersTest.java
@@ -72,10 +72,7 @@ class ExpressionTreeComparersTest {
 
   BslExpression getExpressionTree(String code) {
     var expression = parse(code);
-    var rewriter = new ExpressionParseTreeRewriter();
-    expression.accept(rewriter);
-
-    return rewriter.getExpressionTree();
+    return ExpressionParseTreeRewriter.buildExpressionTree(expression);
   }
 
 }
